desktop: add urgent count tracking and optimize operations
Add urgent_count field to desktop_t to enable O(1) urgent checking
instead of O(n) tree traversal. Batch EWMH updates to reduce X11
round trips.

Security improvements:
- Add null checks at function entry points
- Use strncpy with explicit null termination for buffer safety
- Check calloc return value in make_desktop()
- Fix signed/unsigned comparison warning in rule.c

Performance improvements:
- Replace multiple EWMH calls with batch_ewmh_update()
- Cache urgent window count per desktop
- Simplify control flow and remove redundant conditions
- Reduce function call overhead with inline batch helper

The urgent_count must be maintained by window management operations
(set_urgent, manage_window, unmanage_window, transfer_node) to keep
the count accurate.

+++ src/rule.c
@@ -429,7 +429,8 @@ void parse_rule_consequence(int fd, rule_consequence_t *csq)
 
     while ((nb = read(fd, data + total_read, sizeof(data) - total_read - 1)) > 0) {
         total_read += nb;
-        if (total_read >= sizeof(data) - 1) {
+        if ((size_t)total_read >= sizeof(data) - 1) {
+
             break;
         }
     }
